Add files via upload
Need help with this projject. I cant get the array or list to save a title then jump to text then save it in a list or array. Also im supposed to be able to seatch for blogg title and write it out, and write out all bloggs, like you see in the menu. Im really stressed so i cant think correctly.